requirements/package.txt
setuptools>=30
six
python-dateutil
sympy
jinja2
django-extended-choices
jsonschema
